chore(behavior_path_planner): move debug utils
